Added class library for dealing with JSON input. Tested correct reading and parsing of JSON data

+++ MoviesSystem/MoviesSystem.Models/Rate.cs
@@ -7,6 +7,6 @@ namespace MoviesSystem.Models
         public int Id { get; set; }
 
         [Range(1,10)]
-        public int RateValue { get; set; }
+        public float RateValue { get; set; }
     }
 }
